+++ example/pages/index.vue
@@ -26,6 +26,13 @@
     <nav>
       <link-list :items="bothLinks" critical />
     </nav>
+    <headline tag="h2" critical>
+      Experiments
+    </headline>
+    <nav>
+      <link-list :items="experimentsLinks" critical />
+    </nav>
+    <br>
   </div>
 </template>
 
@@ -39,7 +46,7 @@ export default {
     Headline,
     LinkList
   },
-  data () {
+  asyncData () {
     return {
       links: [
         {
@@ -67,12 +74,12 @@ export default {
           to: '/tests/v-font-media'
         },
         {
-          title: 'Test: LazyPicture',
-          to: '/tests/lazy-picture'
+          title: 'Test: SpeedkitPicture',
+          to: '/tests/speedkit-picture'
         },
         {
-          title: 'Test: LazyImage',
-          to: '/tests/lazy-image'
+          title: 'Test: SpeedkitImage',
+          to: '/tests/speedkit-image'
         },
         {
           title: 'Test: Speedkit Function',
@@ -97,16 +104,16 @@ export default {
           to: '/critical/font-extended'
         },
         {
-          title: 'LazyImage',
-          to: '/critical/lazy-image'
+          title: 'SpeedkitImage',
+          to: '/critical/speedkit-image'
         },
         {
-          title: 'LazyPicture',
-          to: '/critical/lazy-picture'
+          title: 'SpeedkitPicture',
+          to: '/critical/speedkit-picture'
         },
         {
-          title: 'LazyIframe',
-          to: '/critical/lazy-iframe'
+          title: 'SpeedkitIframe',
+          to: '/critical/speedkit-iframe'
         }
       ],
       lazyLinks: [
@@ -119,16 +126,16 @@ export default {
           to: '/lazy/font-extended'
         },
         {
-          title: 'LazyImage',
-          to: '/lazy/lazy-image'
+          title: 'SpeedkitImage',
+          to: '/lazy/speedkit-image'
         },
         {
-          title: 'LazyPicture',
-          to: '/lazy/lazy-picture'
+          title: 'SpeedkitPicture',
+          to: '/lazy/speedkit-picture'
         },
         {
-          title: 'LazyIframe',
-          to: '/lazy/lazy-iframe'
+          title: 'SpeedkitIframe',
+          to: '/lazy/speedkit-iframe'
         }
       ],
       bothLinks: [
@@ -141,18 +148,28 @@ export default {
           to: '/both/font-extended'
         },
         {
-          title: 'LazyImage',
-          to: '/both/lazy-image'
+          title: 'SpeedkitImage',
+          to: '/both/speedkit-image'
         },
         {
-          title: 'LazyPicture',
-          to: '/both/lazy-picture'
+          title: 'SpeedkitPicture',
+          to: '/both/speedkit-picture'
         },
         {
-          title: 'LazyIframe',
-          to: '/both/lazy-iframe'
+          title: 'SpeedkitIframe',
+          to: '/both/speedkit-iframe'
         }
-      ]
+      ],
+      experimentsLinks: [
+        {
+          title: 'SpeedkitYoutube',
+          to: '/experiments/speedkit-youtube'
+        }]
+    }
+  },
+  data () {
+    return {
+      links: []
     }
   }
 }