+++ lib/components/SpeedkitLayer.vue
@@ -12,7 +12,7 @@
 
 <script>
 
-import OnlySsr from './OnlySsr'
+import OnlySsr from './abstracts/OnlySsr'
 export default {
   name: 'SpeedkitInfoLayer',
 
@@ -22,11 +22,11 @@ export default {
   props: {
     hide: {
       type: Boolean,
-      default: <%= options.ignorePerformance %>
+      default: process.env.NUXT_SPEEDKIT_IGNORE_PERFORMANCE
     },
     ignoreNoSsr: {
       type: Boolean,
-      default: <%= options.ignorePerformance || false %>
+      default: process.env.NUXT_SPEEDKIT_IGNORE_PERFORMANCE
     }
   },
   computed: {